Agent/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
edge-tts>=7.2.0
imageio-ffmpeg>=0.4.9
json-repair>=0.30.0
minio>=7.2.0
numpy>=1.26.0
openai>=1.40.0
opencv-python>=4.9.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.34.0
